#38
- preset/isomorphic: добавлена точка входа с типами (minor)
- HandlerContext перенесен в preset/node (major)
