+++ src/preset/node/index.ts
@@ -1,2 +1,3 @@
+export type { HandlerContext } from './types';
 export { PresetNode } from './node';
 export { PresetHandler, HandlerProvider } from './handler';
