+++ src/preset/node/types.ts
@@ -1,0 +1,10 @@
+import type { Request, Response, NextFunction } from 'express';
+
+/**
+ * Контекст обработчика express.
+ */
+export interface HandlerContext {
+  req: Request;
+  res: Response;
+  next: NextFunction;
+}
