+++ src/tokens.ts
@@ -12,12 +12,8 @@ import type { Tracer } from '@opentelemetry/api';
 import type { BasicTracerProvider, SpanExporter } from '@opentelemetry/sdk-trace-base';
 import type { Resource } from '@opentelemetry/resources';
 import type { ElementType, ReactNode } from 'react';
-import type {
-  HandlerContext,
-  KnownHttpApiKey,
-  PageAssets,
-  StrictMap,
-} from './preset/isomorphic/types';
+import type { KnownHttpApiKey, PageAssets, StrictMap } from './preset/isomorphic/types';
+import type { HandlerContext } from './preset/node/types';
 import type { SpecificExtras } from './preset/node/handler/utils';
 import type { CreateAxiosDefaults } from 'axios';
 import type { AxiosInstanceWrapper } from 'middleware-axios';
